Keep plugin install implementation details private
