+++ codex-rs/core/src/tools/request_plugin_installs_tests.rs
@@ -12,16 +12,8 @@ use codex_config::types::ToolSuggestDiscoverableType;
 use codex_core_plugins::PluginInstallRequest;
 use codex_core_plugins::PluginsManager;
 use codex_core_plugins::startup_sync::curated_plugins_repo_path;
-use codex_plugin_installs_extension::MAX_REQUEST_PLUGIN_INSTALLS_ENTRIES;
-use codex_plugin_installs_extension::RequestPluginInstallPickerCategory;
-use codex_plugin_installs_extension::RequestPluginInstallPickerEntry;
-use codex_plugin_installs_extension::RequestPluginInstallsArgs;
-use codex_plugin_installs_extension::ToolSuggestPresentation;
-use codex_plugin_installs_extension::request_plugin_install_picker_completed;
-use codex_plugin_installs_extension::validate_request_plugin_install_picker_args;
 use codex_rmcp_client::ElicitationResponse;
 use codex_tools::DiscoverablePluginInfo;
-use codex_tools::DiscoverableToolAction;
 use codex_utils_absolute_path::AbsolutePathBuf;
 use core_test_support::PathExt;
 use pretty_assertions::assert_eq;
@@ -113,182 +105,6 @@ fn request_plugin_install_response_persists_only_decline_always_mode() {
 }
 
 #[test]
-fn validate_request_plugin_install_picker_args_supports_categories() {
-    let args = RequestPluginInstallsArgs {
-        action_type: DiscoverableToolAction::Install,
-        entries: None,
-        categories: Some(vec![RequestPluginInstallPickerCategory {
-            title: "Calendar".to_string(),
-            entries: vec![RequestPluginInstallPickerEntry {
-                tool_id: "connector_calendar".to_string(),
-                tool_type: DiscoverableToolType::Connector,
-            }],
-        }]),
-    };
-    let discoverable_tools = vec![connector_tool("connector_calendar", "Google Calendar")];
-
-    let resolved_entries = validate_request_plugin_install_picker_args(
-        &args,
-        &discoverable_tools,
-        /*app_server_client_name*/ None,
-        ToolSuggestPresentation::ListTool,
-    )
-    .expect("categorized picker args");
-
-    assert_eq!(resolved_entries.len(), 1);
-    assert_eq!(resolved_entries[0].category_index, Some(0));
-    assert_eq!(resolved_entries[0].tool.id(), "connector_calendar");
-}
-
-#[test]
-fn validate_request_plugin_install_picker_args_rejects_mixed_sources() {
-    let entry = RequestPluginInstallPickerEntry {
-        tool_id: "connector_calendar".to_string(),
-        tool_type: DiscoverableToolType::Connector,
-    };
-    let args = RequestPluginInstallsArgs {
-        action_type: DiscoverableToolAction::Install,
-        entries: Some(vec![entry]),
-        categories: Some(vec![RequestPluginInstallPickerCategory {
-            title: "Calendar".to_string(),
-            entries: vec![RequestPluginInstallPickerEntry {
-                tool_id: "connector_calendar".to_string(),
-                tool_type: DiscoverableToolType::Connector,
-            }],
-        }]),
-    };
-    let discoverable_tools = vec![connector_tool("connector_calendar", "Google Calendar")];
-
-    assert_eq!(
-        validate_request_plugin_install_picker_args(
-            &args,
-            &discoverable_tools,
-            /*app_server_client_name*/ None,
-            ToolSuggestPresentation::ListTool,
-        )
-        .expect_err("mixed picker args"),
-        FunctionCallError::RespondToModel(
-            "picker install requests must include exactly one of entries or categories".to_string(),
-        ),
-    );
-}
-
-#[test]
-fn validate_request_plugin_install_picker_args_rejects_duplicate_tools() {
-    let entry = RequestPluginInstallPickerEntry {
-        tool_id: "connector_calendar".to_string(),
-        tool_type: DiscoverableToolType::Connector,
-    };
-    let args = RequestPluginInstallsArgs {
-        action_type: DiscoverableToolAction::Install,
-        entries: None,
-        categories: Some(vec![
-            RequestPluginInstallPickerCategory {
-                title: "Calendar".to_string(),
-                entries: vec![entry],
-            },
-            RequestPluginInstallPickerCategory {
-                title: "Meetings".to_string(),
-                entries: vec![RequestPluginInstallPickerEntry {
-                    tool_id: "connector_calendar".to_string(),
-                    tool_type: DiscoverableToolType::Connector,
-                }],
-            },
-        ]),
-    };
-    let discoverable_tools = vec![connector_tool("connector_calendar", "Google Calendar")];
-
-    assert_eq!(
-        validate_request_plugin_install_picker_args(
-            &args,
-            &discoverable_tools,
-            /*app_server_client_name*/ None,
-            ToolSuggestPresentation::ListTool,
-        )
-        .expect_err("duplicate picker tool"),
-        FunctionCallError::RespondToModel(
-            "picker install requests must not repeat a tool_type/tool_id pair".to_string(),
-        ),
-    );
-}
-
-#[test]
-fn validate_request_plugin_install_picker_args_rejects_multi_tool_tui_requests() {
-    let args = RequestPluginInstallsArgs {
-        action_type: DiscoverableToolAction::Install,
-        entries: Some(vec![
-            RequestPluginInstallPickerEntry {
-                tool_id: "connector_calendar".to_string(),
-                tool_type: DiscoverableToolType::Connector,
-            },
-            RequestPluginInstallPickerEntry {
-                tool_id: "connector_gmail".to_string(),
-                tool_type: DiscoverableToolType::Connector,
-            },
-        ]),
-        categories: None,
-    };
-    let discoverable_tools = vec![
-        connector_tool("connector_calendar", "Google Calendar"),
-        connector_tool("connector_gmail", "Gmail"),
-    ];
-
-    assert_eq!(
-        validate_request_plugin_install_picker_args(
-            &args,
-            &discoverable_tools,
-            Some("codex-tui"),
-            ToolSuggestPresentation::ListTool,
-        )
-        .expect_err("multi-tool TUI request"),
-        FunctionCallError::RespondToModel(
-            "multi-tool install requests are not available in codex-tui yet".to_string(),
-        ),
-    );
-}
-
-#[test]
-fn validate_request_plugin_install_picker_args_caps_entries() {
-    let entries = || {
-        (0..=MAX_REQUEST_PLUGIN_INSTALLS_ENTRIES)
-            .map(|index| RequestPluginInstallPickerEntry {
-                tool_id: format!("connector_{index}"),
-                tool_type: DiscoverableToolType::Connector,
-            })
-            .collect()
-    };
-
-    for args in [
-        RequestPluginInstallsArgs {
-            action_type: DiscoverableToolAction::Install,
-            entries: Some(entries()),
-            categories: None,
-        },
-        RequestPluginInstallsArgs {
-            action_type: DiscoverableToolAction::Install,
-            entries: None,
-            categories: Some(vec![RequestPluginInstallPickerCategory {
-                title: "Connectors".to_string(),
-                entries: entries(),
-            }]),
-        },
-    ] {
-        assert_eq!(
-            validate_request_plugin_install_picker_args(
-                &args,
-                &[],
-                /*app_server_client_name*/ None,
-                ToolSuggestPresentation::ListTool,
-            )
-            .expect_err("oversized picker args"),
-            FunctionCallError::RespondToModel(format!(
-                "picker install requests support at most {MAX_REQUEST_PLUGIN_INSTALLS_ENTRIES} entries"
-            )),
-        );
-    }
-}
-
-#[test]
 fn picker_response_acknowledgements_apply_only_to_plugins() {
     let connector = RequestedPickerInstallEntry {
         tool: connector_tool("connector_calendar", "Google Calendar"),
@@ -322,26 +138,6 @@ fn picker_response_acknowledgements_apply_only_to_plugins() {
         ),
         (false, true),
     );
-}
-
-#[test]
-fn picker_completion_only_requires_one_completed_entry() {
-    let entries = vec![
-        RequestPluginInstallEntryResult {
-            tool_type: DiscoverableToolType::Connector,
-            tool_id: "connector_salesforce".to_string(),
-            tool_name: "Salesforce".to_string(),
-            completed: true,
-        },
-        RequestPluginInstallEntryResult {
-            tool_type: DiscoverableToolType::Connector,
-            tool_id: "connector_hubspot".to_string(),
-            tool_name: "HubSpot".to_string(),
-            completed: false,
-        },
-    ];
-
-    assert!(request_plugin_install_picker_completed(&entries));
 }
 
 #[tokio::test]
